+++ API Gateway/Dockerfile.dev
@@ -18,7 +18,7 @@ COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
 # Install additional dev dependencies
-RUN pip install --no-cache-dir watchfiles
+RUN pip install --no-cache-dir watchfiles email-validator
 
 # Expose port
 EXPOSE 8080
