+++ API Gateway/requirements.txt
@@ -13,3 +13,6 @@ python-dotenv==1.0.0
 # Kafka and Redis dependencies
 aiokafka==0.10.0
 redis==5.0.1
+
+# Email validation for pydantic
+email-validator==2.1.0
